Update the rest of the examples, prune unused, simplify used
