+++ examples/samples/simple.py
